+++ Comp/common.cpp
@@ -76,7 +76,7 @@ const char* GetTypeName(AComp* AC)
     if (B != 0)
       return _BASHTypeName[B->tip];
   }
-  int ind = _CommonTypeInd[AType];
+  size_t ind = _CommonTypeInd[AType];
   if (ind == _ctnon)
     return Com_Nam[AType];
   return _CommonTypeName[ind];
